refactor(client): DatabaseManager -> ClientDatabaseManager
